+++ example/lib/main.dart
@@ -1,8 +1,13 @@
 import 'package:memescraper/memescraper.dart';
 
-void main() async {
+void main() {
+  listMiniMemesAndMemes();
+  findMemeInfo();
+}
+
+void listMiniMemesAndMemes() async {
   var list = await MemeScraper.listMemes(
-      type: Type.confirmed, sortedBy: SortedBy.newest, page: 2);
+      type: Type.confirmed, sortedBy: SortedBy.oldest, page: 2);
   for (var mini in list) {
     print('Mini : ');
     print(mini);
@@ -11,4 +16,9 @@ void main() async {
     print(await MemeScraper.miniMemeToMeme(mini));
     print('\n');
   }
+}
+
+void findMemeInfo() async {
+  var meme = await MemeScraper.getMemeInfo('rickroll');
+  print(meme);
 }