+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+- Now scrapping all infos about Memes in a map
+
+## 1.0.4
+- Fix order of searching
+
 ## 1.0.3
 - Add possibility to search memes on a precise page